ai/stt/requirements.txt
fastapi
uvicorn
openai-whisper
imageio-ffmpeg
python-multipart
ffmpeg-normalize
requests